ensure surelog marks required keys
