+++ siliconcompiler/tools/surelog/surelog.py
@@ -102,123 +102,3 @@ def parse_version(stdout):
 
     # grab version # by splitting on whitespace
     return stdout.split()[1]
-
-
-################################
-#  Custom runtime options
-################################
-def _remove_dups(chip, type, file_set):
-    new_files = []
-    for f in file_set:
-        if f not in new_files:
-            new_files.append(f)
-        else:
-            chip.logger.warning(f"Removing duplicate '{type}' inputs: {f}")
-    return new_files
-
-
-def runtime_options(chip):
-
-    ''' Custom runtime options, returnst list of command line options.
-    '''
-
-    step = chip.get('arg', 'step')
-    index = chip.get('arg', 'index')
-
-    cmdlist = []
-
-    libs = []
-    libs.extend(chip.get('asic', 'logiclib', step=step, index=index))
-    libs.extend(chip.get('asic', 'macrolib', step=step, index=index))
-
-    #####################
-    # Library directories
-    #####################
-
-    ydir_files = chip.find_files('option', 'ydir')
-
-    for item in libs:
-        ydir_files.extend(chip.find_files('library', item, 'option', 'ydir'))
-
-    # Deduplicated source files
-    for value in _remove_dups(chip, 'ydir', ydir_files):
-        cmdlist.append('-y ' + value)
-
-    #####################
-    # Library files
-    #####################
-
-    vlib_files = chip.find_files('option', 'vlib')
-
-    for item in libs:
-        vlib_files.extend(chip.find_files('library', item, 'option', 'vlib'))
-
-    for value in _remove_dups(chip, 'vlib', vlib_files):
-        cmdlist.append('-v ' + value)
-
-    #####################
-    # Include paths
-    #####################
-
-    idir_files = chip.find_files('option', 'idir')
-
-    for item in libs:
-        idir_files.extend(chip.find_files('library', item, 'option', 'idir'))
-
-    for value in _remove_dups(chip, 'idir', idir_files):
-        cmdlist.append('-I' + value)
-
-    #######################
-    # Variable Definitions
-    #######################
-
-    # Extra environment variable defines (don't need deduplicating)
-    for value in chip.get('option', 'define'):
-        cmdlist.append('-D' + value)
-
-    for item in libs:
-        for value in chip.get('library', item, 'option', 'define'):
-            cmdlist.append('-D' + value)
-
-    #######################
-    # Command files
-    #######################
-
-    # Command-line argument file(s).
-    cmd_files = chip.find_files('option', 'cmdfile')
-
-    for item in libs:
-        cmd_files.extend(chip.find_files('library', item, 'option', 'cmdfile'))
-
-    for value in _remove_dups(chip, 'cmdfile', cmd_files):
-        cmdlist.append('-f ' + value)
-
-    #######################
-    # Sources
-    #######################
-
-    src_files = chip.find_files('input', 'rtl', 'verilog', step=step, index=index)
-
-    # TODO: add back later
-    # for item in libs:
-    #    src_files.extend(chip.find_files('library', item, 'input', 'verilog'))
-
-    for value in _remove_dups(chip, 'source', src_files):
-        cmdlist.append(value)
-
-    #######################
-    # Top Module
-    #######################
-
-    cmdlist.append('-top ' + chip.top())
-
-    ###############################
-    # Parameters (top module only)
-    ###############################
-
-    # Set up user-provided parameters to ensure we elaborate the correct modules
-    for param in chip.getkeys('option', 'param'):
-        value = chip.get('option', 'param', param)
-        cmdlist.append(f'-P{param}={value}')
-
-    return cmdlist
